rename maincategory to portfolio
